fix: unmarked session when restarting
When session is restarted because of heart beat lapsed we close the
current session and opens up a new session and the internal mapping with
the browserid in the provider is also changed. So this leads to one
unmarked session and can be many depending upon the count.

Before closing the browser, we can basically introduce a new job status
as restarted so that users are aware that why there are more session
than the desired count. Fix is to call the reportJob result before
closing the browser.