add: not terminating process on one connection
When there is an error in one of the browser-connection due to any error
reason, the complete process is closed because the error event promise
is resolved and thus disposing of assets gets started.

What actually should be done is perform the cleanup process when the
error is being emitted from all the connection and then emit the browset
error and do not persuade the cleanup on the pending tasks. Logic for
this would be similar to follows.

```
connection.on('error', () => {
        set.insert(connection);
        if (set.size === total_connection)
            this.emit('error'); // This will initiate disposing
});
```