perf(compiler/babel): load configuration only once
Loading the Babel configuration is an expensive operation. Since the configuration file doesn't change during the compilation it should only be read once. Currently it is read each time a client function is transpiled (in `downgradeES`). Profiling has shown that for small selectors the loading of the configuration file takes significantly (up to 5 times) more than the transpilation of the code.

This patch introduces caching of the configuration file, resulting in a ~33% reduction of the duration of the compilation for certain use cases and lower memory usage. While the addition of the caching is a bit ad-hoc (as opposed to a more systematic approach where a caching layer is introduced), we consider that the performance improvement (which can easily translate to tens of seconds for large test suites) is worth the small decrease in code readability (and consistency).

Closes https://github.com/DevExpress/testcafe/issues/6284.