Proxyless context storage (#7467)
related PR in HH:
https://github.com/DevExpress/testcafe-hammerhead/pull/2841
review HH firstly

**What was done:**

- revert cdp skip-js-errors
- revert cdp native-dialogs
- revert cdp console-messages
- add server-side proxyless contextStorage

Implement proxyless strategy for storing contextStorage information. 
Previously it was window.sessionStorage. Now this information is stored
on the server side.
The synchronization from client to server is made using the
`Runtime.addBinding` method:
https://chromedevtools.github.io/devtools-protocol/tot/Runtime/#method-addBinding
Initially I wanted to synchronize only on the `beforeunload` event, but
it did not work for iframes and some other edge cases. So, now I
synchronize the data immediately.

On the server side we keep this information inside the
ProxylessRequestPipelineContext and restore it every time when we need
it using ResourceInjector.

Also, I change the way of how we restore storages for roles, since the
previous approach does not work anymore, because we do not override
storages anymore.

Also, I clear storages on each testRun using the
`client.Storage.clearDataForOrigin` function.