Basic Http-auth in proxyless mode (#7476)
- only HTTP Basic auth
- Enter credentials only on failed request with 401 status (as in HH
version)

**why?**
we cannot use
[Fetch.authRequired](https://chromedevtools.github.io/devtools-protocol/tot/Fetch/#event-authRequired)
event, since in this case Chrome will store credentials inside the user
profile. We cannot reset credentials in this case without reopening
browser.
Moreover, we cannot sent Authorization header on every request, since in
this case other target websites can get access to the Authorization
header + this leads to unexpected errors in browser (but I did not
research these problems).

For making request, I use the method which was used in the Request API.
We handle the `requestPaused` event on the response stage. We check
responseStatus. If responseStatus === 401, we are making the same
request (but additionally we provide auth credentials) to the target
website. Then we fulfill the original request with data from the
additional request.