feat: selector inspector panel (closes #7375) (#7372)
Closes #7375.

### Use cases

1. **Picking + generation**. Click on the `Pick` button and select the
element with a mouse click. As a result, several selectors matching this
element will be generated. The first of the generated selectors will be
placed in the selector input, the rest are available in the list of
selectors, which can be expanded by clicking on the arrow to the left of
the `Copy` button.
2. **Debug selector**. Start typing the selector in CSS or TestCafe
format. As you type, all matching elements will be highlighted with a
frame. In addition, to the right of the input field there is an
indicator that shows the number of matching elements.

### Peculiarities

**In the picking mode** all elements of TestCafe UI, except for the
frame that highlights the element, disappear.

**Input selector format**. The Selector Inspector recognizes selectors
in two formats: CSS and TestCafe.
CSS: `div > div input`, `p:nth-child(2)`,
`span[custom-attribute="value"].className`
TestCafe: `Selector('.my-class')`, `Selector('div').withText('Hello
World!')`

**Match indicator**. It has three states: `not found`, `incorrect
selector` and `found: {number of elements}`.

**List of generated selectors**. Cleared as soon as the user starts
typing in the selector input field.

**Highlighting elements** matching the input selector is active only
when the selector input has focus.

### Restrictions

1. The inspector is currently only available in debug mode (`await
t.debug()`).
2. Picking does not work in iframes.
3. Picking doesn't work with Shadow DOM.
4. The selector generator is not configurable.
5. There is only one generator so far, and the user will not be able to
redefine the generation function.

## Initial state

![изображение](https://user-images.githubusercontent.com/34184692/213652793-ad8cc93f-8f68-466d-870f-28a89036d519.png)

## Picking

![021](https://user-images.githubusercontent.com/34184692/213653402-c08f2723-7343-402f-9b2f-2b2a5c0e3134.gif)

## Selector typing

![022](https://user-images.githubusercontent.com/34184692/213653738-ff59cb80-ab19-4a2d-82f9-168955f6a525.gif)

## Selecting a selector from a list

![023](https://user-images.githubusercontent.com/34184692/213656937-ba08fddc-5bba-4fdd-9f55-7d40f37a179a.gif)

## Copy selector

![024](https://user-images.githubusercontent.com/34184692/213654567-77c5ef12-bfa2-4c74-a057-5ecc9b2a6f99.gif)