fix: use calculated osInfo in userAgent (#7600)
<!--
Thank you for your contribution.

Before making a PR, please read our contributing guidelines at

https://github.com/DevExpress/testcafe/blob/master/CONTRIBUTING.md#code-contribution

We recommend creating a *draft* PR, so that you can mark it as 'ready
for review' when you are done.
-->

## Purpose
https://github.com/DevExpress/testcafe-private/issues/147

Both the browserConnection.connectionInfo and
browserConnection.userAgent are prettyUserAgent string. The only
difference is in osInfo. In the case of userAgent it is calculated from
the browser and usually it is incorrect.

## Approach
Calculate osInfo in userAgent using the provider getOSInfo method. 

## References
https://github.com/DevExpress/testcafe-private/issues/147

## Pre-Merge TODO
- [ ] Write tests for your proposed changes
- [ ] Make sure that existing tests do not fail