fix: takeOnFails screenshot with fullPage option fixed (#8077)
<!--
Thank you for your contribution.

Before making a PR, please read our contributing guidelines at

https://github.com/DevExpress/testcafe/blob/master/CONTRIBUTING.md#code-contribution

We recommend creating a *draft* PR, so that you can mark it as 'ready
for review' when you are done.
-->

## Purpose
There is an option for testcaferc configuration file called takeOnFails.
The issue here was that the screenshot taken by takeOnFails was not
responding to option fullPage. The screenshot was always small.

## Approach
I found the place where the fullPage option was not passing. Fixed it by
passing fullPage option to TakeScreenshotOnFailCommand.

## References
https://github.com/DevExpress/testcafe/issues/7761

## Pre-Merge TODO
- [x] Write tests for your proposed changes
- [x] Make sure that existing tests do not fail