feat: debug with selector option added (#8066)
<!--
Thank you for your contribution.

Before making a PR, please read our contributing guidelines at

https://github.com/DevExpress/testcafe/blob/master/CONTRIBUTING.md#code-contribution

We recommend creating a *draft* PR, so that you can mark it as 'ready
for review' when you are done.
-->

## Purpose
The purpose of this PR is to allow passing selector to a t.debug call.
Something like this t.debug(Selector). Then the selector will be shown
in debug mode and highlighted.

## Approach
So overall, the main goal was to pass the selector argument from
server-side debug call to client-side selector-input-contaner.
1) src/test-run/index.ts Inside this file we have
_internalExecuteCommand method. There is debug command type handling
block. Inside of it we call _enqueueSetBreakpointCommand to which we
pass command.selector as a second argument. This way we transfer
Selector from debugCommand to breakpointCommand.
2) After the breakpointCommand is added to driverTaskQueue we can work
with it on client-side.
3) On client-side we pass selector to _debugSelector method of
SelectorInputContainer class.
4) In _debugSelector method we:
           -   remove any previously highlighted elements,
- get Selector name and assign it to this.value for
SelectorInputContainer,
- execute Selector to get elements for highlighting and indication,
           -  highlight and indicate elements of the Selector

## References
https://github.com/DevExpress/testcafe-private/issues/163

## Pre-Merge TODO
- [x] Write tests for your proposed changes
- [x] Make sure that existing tests do not fail