fix: remove unused package js-yaml (#8472)
<!--
Thank you for your contribution.

Before making a PR, please read our contributing guidelines at

https://github.com/DevExpress/testcafe/blob/master/CONTRIBUTING.md#code-contribution

We recommend creating a *draft* PR, so that you can mark it as 'ready
for review' when you are done.
-->

## Purpose
Remove unused package js-yaml, introduced here:
https://github.com/DevExpress/testcafe/commit/aaf14cf895eefcc20c31a011425477715a3bab9a

Currently is not being used in the project.

## Approach
_Describe how your changes address the issue or implement the desired
functionality in as much detail as possible._

## References
_Provide a link to the existing issue(s), if any._

## Pre-Merge TODO
- [ ] Write tests for your proposed changes
- [ ] Make sure that existing tests do not fail

Co-authored-by: adil.rakhaliyev <adil.rakhaliyev@devexpress.com>